requirements.txt
pandas==2.2.0
scipy==1.12.0
streamlit==1.31.1
plotly-express==0.4.0
altair==5.2.0